+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - support iframe in text formatting
 - fix bug where polygon export were adding a point
 - make that only visible elements are downloaded
+- iframe export helper
 
 ## 0.5.x
 - datalayers are now sent to backend as geojson